+++ lib/shopify-cli/context.rb
@@ -329,14 +329,18 @@ module ShopifyCli
     # - `**kwargs`: additional keyword arguments to pass to Process.spawn
     #
     # #### Returns
-    # - `status`: boolean success status of the command execution
+    # - `status`: The `Process::Status` result of the command execution.
     #
     # #### Usage
     #
     #   stat = @ctx.system('ls', 'a_folder')
     #
     def system(*args, **kwargs)
-      CLI::Kit::System.system(*args, env: @env, **kwargs)
+      process_status = CLI::Kit::System.system(*args, env: @env, **kwargs)
+      unless process_status.success?
+        abort("System call failed: #{args.join(' ')}")
+      end
+      process_status
     end
 
     # Execute a command in the user's environment